[VT2-273] refactor: 찜 기능 서비스 레이어 분리
