+++ src/main/java/eureca/capstone/project/orchestrator/transaction_feed/service/impl/TransactionFeedServiceImpl.java
@@ -318,42 +318,6 @@ public class TransactionFeedServiceImpl implements TransactionFeedService {
     }
 
     @Override
-    @Transactional
-    public void addWishFeed(String email, AddWishFeedRequestDto requestDto) {
-        User user = findUserByEmail(email);
-        TransactionFeed transactionFeed = findTransactionFeedById(requestDto.getTransactionFeedId());
-        log.info("[addWishFeed] 사용자 및 판매글 조회 완료.");
-
-        if (transactionFeed.isDeleted()) throw new TransactionFeedNotFoundException();
-        if (likedRepository.existsByFeedAndUser(transactionFeed, user)) throw new InternalServerException(ErrorCode.ALREADY_EXISTS_LIKED_LIST);
-        log.info("[addWishFeed] 찜 목록에 존재 X.");
-
-        Liked liked = Liked.builder()
-                .user(user)
-                .transactionFeed(transactionFeed)
-                .build();
-        likedRepository.save(liked);
-        log.info("[addWishFeed] 찜 목록에 추가 완료. 사용자: {}, 판매글: {}", user.getUserId(), transactionFeed.getTransactionFeedId());
-    }
-
-    @Override
-    @Transactional
-    public void removeWishFeed(String email, RemoveWishFeedsRequestDto requestDto) {
-        List<Long> transactionFeedIds = requestDto.getTransactionFeedIds();
-
-        if (CollectionUtils.isEmpty(transactionFeedIds)) {
-            log.info("[removeWishFeed] 삭제할 ID 리스트가 비어있어 작업을 종료합니다.");
-            return;
-        }
-
-        User user = findUserByEmail(email);
-        log.info("[removeWishFeed] 사용자 {}의 찜 목록 삭제 시작. 대상 판매글 수: {}", user.getUserId(), transactionFeedIds.size());
-
-        likedRepository.removeByUserAndFeedIds(transactionFeedIds, user);
-        log.info("[removeWishFeed] 찜 목록에서 삭제 완료. 사용자: {}, 대상 판매글: {}", user.getUserId(), transactionFeedIds);
-    }
-
-    @Override
     @Transactional(readOnly = true)
     public long reindexAllFeeds() {
         // 1. 기존 Elasticsearch 인덱스 삭제
